+++ src/main/java/com/bwd/bwd/request/JobsmithReportRequest.java
@@ -17,7 +17,7 @@ public class JobsmithReportRequest {
 	public String jobsmith_report_name;
 	public String jobsmith_report_note;
 	public int companyid;
-	private long useraccountid;
+	private int useraccountid;
 	private String userid;
 	
 	public ArrayList<Critical> critical;
